Separated logic from Screen
Created hook for login screen state to separate the logic from the view
