+++ renta_facil_colima/src/View/Screens/LoginScreen.jsx
@@ -5,40 +5,24 @@ import GoogleButton from '../buttons/GoogleButton';
 import Loginstyles from '../../styles/LoginScreen';
 import LogoRFC from '../components/LogoRFC';
 import GoogleSignUp from '../../hooks/GoogleSingUp';
+import UseLoginScreenState from '../../hooks/UseLoginScreenState';
 const LoginScreen = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [empityEmail, setValueEmail] = useState(false);
-  const [empityPassword, setValuePassword] = useState(false);
-
-  const isEmpity = () => {
-    setValueEmail(email.trim() === '');
-    setValuePassword(password.trim() === '');
-
-    if (email.trim() === '' || password.trim() === '') {
-      Alert.alert('Algun campo esta vacio');
-    } else {
-      validateEmail();
-    }
-  };
-
-  const validateEmail = () => {
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-
-    if (emailRegex.test(email)) {
-      Alert.alert('Email válido', 'El email es válido.');
-    } else {
-      Alert.alert('Email inválido', 'Por favor, ingresa un email válido.');
-    }
-  };
-
+  const {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    emptyEmail,
+    emptyPassword,
+    isEmpty,
+  } = UseLoginScreenState();
   return (
     <View style={Loginstyles.container}>
       <View style={Loginstyles.logoContainer}>
         <LogoRFC showAppName={false} />
       </View>
       <View
-        style={[Loginstyles.shadow, empityEmail ? {borderColor: 'red'} : null]}>
+        style={[Loginstyles.shadow, emptyEmail ? {borderColor: 'red'} : null]}>
         <TextInput
           style={Loginstyles.inputs}
           onChangeText={text => setEmail(text)}
@@ -52,7 +36,7 @@ const LoginScreen = () => {
       <View
         style={[
           Loginstyles.shadow,
-          empityPassword ? {borderColor: 'red'} : null,
+          emptyPassword ? {borderColor: 'red'} : null,
         ]}>
         <TextInput
           style={Loginstyles.inputs}
@@ -66,7 +50,7 @@ const LoginScreen = () => {
       <View style={Loginstyles.buttonsContainer}>
         <BasicButton
           text="Login"
-          onPress={isEmpity}
+          onPress={isEmpty}
           textSize={32}
           borderColor="#058C42"
           backgroundColor="#FFF"
@@ -76,7 +60,6 @@ const LoginScreen = () => {
           width={308}
           shadow={true}
         />
-
         <GoogleButton onPress={GoogleSignUp}/>
       </View>
     </View>
